Moved UpdatePost_TitleIsUpdated_UpdateSlug() test to EditTests
